requirements.txt


# external requirements
click==8.1.7

flake8==7.0.0
Flask==3.0.0
numpy==1.24.3
opencv-python==4.9.0.80
python-dotenv==1.0.0
SciPy==1.11.4
tensorflow==2.13
dvc
opendatasets
# local package
-e .